+++ backend/controllers/movie.controller.js
@@ -1,12 +1,18 @@
 const baseResponse = require('../dto/baseResponse');
 const movieService = require('../services/movie.service');
 const {StatusCodes} = require('http-status-codes');
-
+const helpers = require('../utils/helpers');
 
 exports.createMovie = async (req, res) => {
  try {
-  //invalid request
-
+  const isInvalid = helpers.handleValidationErrors(req);
+  if (isInvalid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInvalid
+   });
+   return
+  }
   const json = await movieService.createMovie(req, res);
   res.status(StatusCodes.CREATED).json({
    ...baseResponse,
@@ -17,8 +23,7 @@ exports.createMovie = async (req, res) => {
   });
 
  } catch (error) {
-  console.log(error);
-//log to error
+  helpers.logToError(error, req, 'Error in Create Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -43,7 +48,7 @@ exports.getAllMovies = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Get All Movies')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -59,7 +64,14 @@ exports.getAllMovies = async (req, res) => {
 
 exports.getAllMoviesWithPagination = async (req, res) => {
  try {
-  //invalid control here
+  const isInvalid = helpers.handleValidationErrors(req);
+  if (isInvalid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInvalid
+   });
+   return
+  }
   const json = await movieService.getAllMoviesWithPagination(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -69,7 +81,7 @@ exports.getAllMoviesWithPagination = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Get All Movies With Pagination')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -85,7 +97,14 @@ exports.getAllMoviesWithPagination = async (req, res) => {
 
 exports.getById = async (req, res) => {
  try {
-  //invalid control here
+  const isInvalid = helpers.handleValidationErrors(req);
+  if (isInvalid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInvalid
+   });
+   return
+  }
   const json = await movieService.getById(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -95,6 +114,7 @@ exports.getById = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
+  helpers.logToError(error, req, 'Error in Get Movie By Id')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -109,7 +129,14 @@ exports.getById = async (req, res) => {
 
 exports.updateMovieById = async (req, res) => {
  try {
-  //invalid control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if ( isInValid ) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   });
+   return
+  }
   const json = await movieService.updateMovieById(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -120,7 +147,7 @@ exports.updateMovieById = async (req, res) => {
   });
 
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in Update Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -135,7 +162,14 @@ exports.updateMovieById = async (req, res) => {
 
 exports.deleteMovieById = async (req, res) => {
  try {
-  //invalid control here
+  const isInvalid = helpers.handleValidationErrors(req);
+  if (isInvalid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInvalid
+   })
+   return
+  }
   const json = await movieService.deleteMovieById(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -145,7 +179,7 @@ exports.deleteMovieById = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in Delete Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -159,7 +193,14 @@ exports.deleteMovieById = async (req, res) => {
 
 exports.uploadImage = async (req, res) => {
  try {
- //validation control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if(isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   });
+   return
+  }
   const json = await movieService.uploadImage(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -169,7 +210,7 @@ exports.uploadImage = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
- //log to error
+  helpers.logToError(error,req,'Error in image upload Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -184,7 +225,14 @@ exports.uploadImage = async (req, res) => {
 
 exports.updateImage = async (req, res) => {
  try {
- //validation control here
+  const isInValid = helpers.handleValidationErrors(req)
+  if (isInValid) {
+   res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   });
+   return
+  }
   const json = await movieService.updateImage(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -194,7 +242,7 @@ exports.updateImage = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
- //log to error
+  helpers.logToError(error,req,'Error in updateImage Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -207,11 +255,16 @@ exports.updateImage = async (req, res) => {
 }
 
 
-
-
 exports.updateScore = async (req, res) => {
  try {
-  //validation control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if(isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   });
+   return
+  }
   const json = await movieService.updateScore(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -221,7 +274,7 @@ exports.updateScore = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in updateScore Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -235,7 +288,14 @@ exports.updateScore = async (req, res) => {
 
 exports.getScore = async (req, res) => {
  try {
-  //validation control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if( isInValid){
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   });
+   return
+  }
   const json = await movieService.getScore(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -245,7 +305,7 @@ exports.getScore = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in getScore Movie')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
